+++ fleet_management/pages/api/vehicles/[veh_id]/servicing/[svc_id]/index.js
@@ -2,7 +2,9 @@ import pool from '../../../../db/index';
 
 export default async function handler(req, res) {
   let method = req.method;
-  let { veh_id, svc_id, servicedate, description, cost } = req.query;
+  let { veh_id, svc_id, servicedate, description, cost, role } = req.query;
+
+  console.log(role, 'from svc api');
 
   if (req.method === 'GET') {
     try {
@@ -31,16 +33,20 @@ export default async function handler(req, res) {
   }
 
   if (req.method === 'DELETE') {
-    try {
-      const data = await pool.query(
-        'DELETE FROM servicing WHERE veh_id = $1 AND id = $2',
-        [veh_id, svc_id]
-      );
-      res.status(200).json({
-        message: `Successfully deleted servicing ${svc_id} for vehicle ${veh_id}`,
-      });
-    } catch (err) {
-      res.status(500).json({ error: 'failed to delete data' });
+    if (role && role === 'admin') {
+      try {
+        const data = await pool.query(
+          'DELETE FROM servicing WHERE veh_id = $1 AND id = $2',
+          [veh_id, svc_id]
+        );
+        res.status(200).json({
+          message: `Successfully deleted servicing ${svc_id} for vehicle ${veh_id}`,
+        });
+      } catch (err) {
+        res.status(500).json({ error: 'failed to delete data' });
+      }
+    } else {
+      res.status(403).json({ error: 'Unauthorized' });
     }
   }
 }
